teleprompter: graduate out of prototype
Crawl is the reading model that won; the stepper and band variants are gone
along with the variant switcher and the tuning popup. Renamed everything off
`*-prototype`, so this is now app/features/teleprompter + /teleprompter.

- Type is fixed in TYPE (26px, 1.55, 25ch, warm) — tuned against the panel.
- Speed stays adjustable, and is calibrated from the prose height only.
  Headings and cues no longer count as spoken words, which was dragging the
  crawl below the wpm on the dial. Range widened to 80-400, default 200.
- Mouse wheel scrolls the script directly.
- Dropped the mirror toggle: Camera Hub mirrors the display itself.
- vite: allow serving a symlinked node_modules, so dev in a git worktree
  doesn't 403 the client entry and silently skip hydration.
- Split the two deferred fs panels out of video-player-panel.tsx, which was
  already over the file-token limit before this branch touched it.
