+++ app/features/video-editor/video-editor.tsx
@@ -35,8 +35,8 @@ import {
   useState,
 } from "react";
 import { enableVideoEditorMode } from "@/lib/diagram-window";
-import { enableTeleprompterEditorMode } from "@/lib/teleprompter-prototype-window";
-import type { CaptureStatus } from "@/lib/teleprompter-prototype-protocol";
+import { enableTeleprompterEditorMode } from "@/lib/teleprompter-window";
+import type { CaptureStatus } from "@/lib/teleprompter-protocol";
 import { useFetcher, useRevalidator, useSubmit } from "react-router";
 import type {
   DatabaseId,
@@ -170,7 +170,7 @@ export const VideoEditor = (props: {
 
   useEffect(() => enableVideoEditorMode(), []);
 
-  // PROTOTYPE (teleprompter) — the popup has no picker, so this is the only way
+  // The teleprompter popup has no picker, so this is the only way
   // it learns which video to show and what capture is doing. Read through a ref
   // at pong time so the fast-changing speech-detector state doesn't re-subscribe.
   const teleprompterStateRef = useRef<{
